add the route for delete the user

+++ backend/src/controller/deleteUser.ts
@@ -3,46 +3,39 @@ import { withAccelerate } from "@prisma/extension-accelerate";
 import apiresponse from "../utils/apiResopnse";
 
 async function deleteUser(c: any) {
-  const password: string = await c.req.json();
-  //initialize the prisma client
+  const body = await c.req.json();
+
+  if (!body || !body.password) {
+    c.status(400);
+    return c.json(new apiresponse(false, 400, "Password must need"));
+  }
+
+  const password = body.password;
 
   const prisma = new PrismaClient({
     datasourceUrl: c.env.DATABASE_URL,
   }).$extends(withAccelerate());
 
   try {
-    const user = prisma.user.delete({
-        where:{
+    const user = await prisma.user.delete({
+      where: {
+        id: c.get("userId"),
+        email: c.get("email"),
+        password: password,
+      },
+    });
 
-            id:c.get("userId"),
-            email: c.get("email"),
-            password: password
-        }
-    })
-
-    if(!user){
-        c.status(400)
-        return c.json(
-            new apiresponse(false,400,"Password is incorrect")
-        )
+    if (!user) {
+      c.status(400);
+      return c.json(new apiresponse(false, 400, "Password is incorrect"));
     }
 
-    c.status(200)
-    return c.json(
-        new apiresponse(true,200,"Sucessfully delete the user")
-    )
-    
+    c.status(200);
+    return c.json(new apiresponse(true, 200, "Sucessfully delete the user"));
+
   } catch (error) {
-    c.status(400)
-        return c.json(
-            new apiresponse(false,400,"Password is incorrect")
-        )
+    c.status(400);
+    return c.json(new apiresponse(false, 400, "Password is incorrect"));
   }
-
-
 }
 export default deleteUser;
-//check the user in authorized or not
-//match the password
-//if matched then delte the user
-//if not send password is incorect
